+++ packages/core/framework/src/http/types.ts
@@ -1,5 +1,10 @@
+import type {
+  ZodNullable,
+  ZodObject,
+  ZodOptional,
+  ZodRawShape,
+} from "@medusajs/deps/zod"
 import type { NextFunction, Request, Response } from "express"
-import type { ZodNullable, ZodObject, ZodOptional, ZodRawShape } from "zod"
 
 import {
   FindConfig,
@@ -7,6 +12,7 @@ import {
   RequestQueryFields,
 } from "@medusajs/types"
 import { MedusaContainer } from "../container"
+import { PolicyAction } from "./middlewares/check-permissions"
 import { RestrictedFields } from "./utils/restricted-fields"
 
 /**
@@ -62,6 +68,10 @@ export type MiddlewareRoute = {
   bodyParser?: ParserConfig
   additionalDataValidator?: ZodRawShape
   middlewares?: MiddlewareFunction[]
+  /** @ignore */
+  policies?:
+    | { resource: string; operation: string }
+    | Array<{ resource: string; operation: string | string[] }>
 }
 
 export type MiddlewaresConfig = {
@@ -95,6 +105,9 @@ export type MiddlewareDescriptor = {
   matcher: string
   methods?: MiddlewareVerb | MiddlewareVerb[]
   handler: MiddlewareFunction
+  policies?:
+    | { resource: string; operation: string }
+    | Array<{ resource: string; operation: string | string[] }>
 }
 
 export type BodyParserConfigRoute = {
@@ -187,7 +200,7 @@ export interface MedusaRequest<
   /**
    * The locale for the current request, resolved from:
    * 1. Query parameter `?locale=`
-   * 2. Content-Language header
+   * 2. x-medusa-locale header
    * 3. Store's default locale
    */
   locale?: string
@@ -198,6 +211,7 @@ export interface AuthContext {
   actor_type: string
   auth_identity_id: string
   app_metadata: Record<string, unknown>
+  user_metadata: Record<string, unknown>
 }
 
 export interface PublishableKeyContext {
@@ -211,6 +225,7 @@ export interface AuthenticatedMedusaRequest<
 > extends MedusaRequest<Body, QueryFields> {
   auth_context: AuthContext
   publishable_key_context?: PublishableKeyContext
+  policies?: PolicyAction[]
 }
 
 export interface MedusaStoreRequest<
@@ -219,6 +234,7 @@ export interface MedusaStoreRequest<
 > extends MedusaRequest<Body, QueryFields> {
   auth_context?: AuthContext
   publishable_key_context: PublishableKeyContext
+  policies?: PolicyAction | PolicyAction[]
 }
 
 export type MedusaResponse<Body = unknown> = Response<Body>